+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-### 0.12.1 - Unreleased
+### 0.12.2 - Unreleased
+
+### 0.12.1 (2020-Jul-21)
+
+* Support `includes` in the JSON input for pipeline export. Related to: https://github.com/gocd/gocd/pull/8266
 
 * Fix for pipeline export with multiple materials without name, #144